Support decimal ratings (e.g. 5.5) for movie watches
Changed the rating column from unsignedTinyInteger to decimal(3,1), updated the validation rule from integer to numeric, and changed the DTO property type from int to float.

+++ app/Actions/Media/Movies/MovieWatchData.php
@@ -13,7 +13,7 @@ final readonly class MovieWatchData
         public ?Carbon $watchedAt,
         public bool $yearOnly,
         public ?string $notes,
-        public ?int $rating,
+        public ?float $rating,
     ) {}
 
     public static function fromRequest(MarkMovieWatchedRequest $request): self
@@ -24,7 +24,7 @@ final readonly class MovieWatchData
                 : null,
             yearOnly: (bool) $request->validated('year_only', false),
             notes: $request->validated('notes'),
-            rating: $request->validated('rating') !== null ? (int) $request->validated('rating') : null,
+            rating: $request->validated('rating') !== null ? (float) $request->validated('rating') : null,
         );
     }
 }
